Added the connection between User and Order with REST

+++ Domain/Order.cs
@@ -1,12 +1,11 @@
-﻿
-namespace Domain
+﻿namespace Domain
 {
     public class Order
     {
         public int Id { get; set; }
         public DateTime CreatedAt { get; set; }
         public int UserId { get; set; }
-        public decimal TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public string ShippingAddress { get; set; }
     }
 }
