+++ projects/Order/OrderApplication/DTO/UpdateOrderDTO.cs
@@ -10,7 +10,7 @@ namespace OrderApplication.DTO
     {
         public int Id { get; set; }
         public int UserId { get; set; }
-        public decimal TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public string ShippingAddress { get; set; }
     }
 }
